Header 구조 개선
기존 Navbar에 있는 기능을 Header로 정리,배치를 하였음!
