+++ src/components/Header.css
@@ -42,4 +42,4 @@ nav li {
     margin-right: 20px; /* 로그인 및 회원가입 링크 사이의 간격을 조정 */
 }
 
-/* 다른 Header 스타일 추가 */
+
