+++ src/layouts/Layout.js
@@ -1,12 +1,12 @@
 import Header from "../components/Header";
-import Navbar from "../components/Navbar";
+// import Navbar from "../components/Navbar";
 import { Outlet } from "react-router-dom";
 
 function Layout() {
     return(
         <>
             <Header/>
-            <Navbar/>
+            {/* <Navbar/> */}
             <Outlet/>
             {/* footer */}
         </>
